Fail Skinning attempts if the player isn't within loot range
On large corpses with disjointed models (i.e. Devilsaurs with their
necks and tail sticking out), the client can think it's in range to
skin without also being within the INTERACTION_DISTANCE required to
loot.  If this occurs, the SendLoot() call fails, no loot is
retreived, and the corpse is no longer skinnable.

Make sure the player is within looting range of the corpse they're
skinning, and if not, report that the spell was out of range.